exercise: reestructuracion de clases y relaciones

+++ src/main/java/org/iesvdm/ejercicios/ejercicio5/auxClasses/Real.java
@@ -7,8 +7,8 @@ public class Real {
     private BigDecimal value;
 
     //Constructores
-    public Real(double value) {
-        this.value = BigDecimal.valueOf(value);
+    public Real(BigDecimal value) {
+        this.value = value;
     }
 
     public Real(String value) {
